3: Implemented fuzzy search.

+++ src/commands/plaintext_command_output_processor.rs
@@ -28,17 +28,13 @@ const INDENT_STEP: &str = "  ";
 
 
 /// Prints the given output variant to the specified destination in a plaintext format with given indentation level.
-/// - If varian is just a value, it is printed as is and nothing else.
+/// - If variant is just a value, it is printed as is and nothing else.
 /// - If variant is an array, each element is printed on a new line with increased indentation
 /// - If variant is a dictionary, each key-value pair is printed on a new line with increased indentation for values
-
-
 fn print(dst: &mut dyn std::io::Write, output: &CommandOutputVariant, level: usize) {
     match output {
         CommandOutputVariant::Value(val) => {
-            let indent = INDENT_STEP.repeat(level as usize);
-
-
+            let indent = INDENT_STEP.repeat(level);
             let _ = write!(dst, "{}{}", indent, val);
         }
         CommandOutputVariant::Array(arr) => {
@@ -48,17 +44,13 @@ fn print(dst: &mut dyn std::io::Write, output: &CommandOutputVariant, level: usi
         }
         CommandOutputVariant::Dict(dict) => {
             for (key, value) in dict {
-                let indent = INDENT_STEP.repeat(level as usize);
-
+                let indent = INDENT_STEP.repeat(level);
 
                 let _ = writeln!(dst, "{}{}", indent, key);
-
-
                 print(dst, value, level + 1);
             }
         }
     }
-
 
     let _ = write!(dst, "\n");
 }
@@ -66,7 +58,7 @@ fn print(dst: &mut dyn std::io::Write, output: &CommandOutputVariant, level: usi
 
 impl CommandOutputProcessor for PlaintextCommandOutputAndErrorProcessor {
     fn process_output(&mut self, output: &CommandOutputVariant) {
-        print(&mut *self.cout, output, 0);
+        print(&mut *self.cout, output, 0usize);
     }
 }
 
@@ -75,17 +67,11 @@ impl CommandErrorProcessor for PlaintextCommandOutputAndErrorProcessor {
     fn process_error(&mut self, error: &anyhow::Error) {
         spdlog::error!("Error: {}", error);
 
-
         let mut source = error.source();
-
-
         while let Some(inner) = source {
             spdlog::error!("Caused by: {}", inner);
-
-
             source = inner.source();
         }
-
 
         let _ = writeln!(self.cerr, "Error: {}", error);
     }
